*Fix issue with the progress bar. Forgot to use float division to calculate progress

+++ src/main/org/testng/eclipse/ui/ProgressBar.java
@@ -92,7 +92,7 @@ public class ProgressBar extends Canvas {
   }
 
   private int calcBarWidth(Rectangle rect){
-    int methodsTotalCount = scoreBoard.getTotalMethodCount();
+    float methodsTotalCount = scoreBoard.getTotalMethodCount();
     if(methodsTotalCount>0){
       /*
       //float progress;
